Move CHUNK_InGameUI parsing to more appropriate place
